+++ Clase 07/Clase 07/array.h
@@ -1,7 +1,7 @@
 #ifndef ARRAY_H_INCLUDED
 #define ARRAY_H_INCLUDED
 
-int utn_mostrarArray(int* pArray,int limite);
+int utn_mostrarArray(float* pArray,int limite);
 int utn_inicializarArray(int* pArray,int valor, int limite);
 int utn_calcularMaximoArray(int* pArray,int limite,int* pMaximo);
 int utn_calcularMinimoArray(int* pArray,int limite,int* pMinimo);
